Create pH_test.ino
The file utilized for pH testing on the Arduino IoT platform integrates sensors to measure the pH levels of aqueous solutions. It employs an analog-to-digital converter (ADC) to interpret voltage readings from the pH sensor, processes the data using Arduino's firmware, and transmits the results to cloud-based IoT applications for monitoring.